LayoutTests/fast/js/kde/encode_decode_uri-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL checkWithFunctions(encodeURI,decodeURI) should be true (of type boolean). Was false (of type boolean).
FAIL checkWithFunctions(encodeURIComponent,decodeURIComponent) should be true (of type boolean). Was false (of type boolean).
PASS successfullyParsed is true

TEST COMPLETE
